adding VO and utils

+++ src/main/java/com/randreucetti/peoplebase/dao/PersonDao.java
@@ -7,7 +7,7 @@ import com.randreucetti.peoplebase.model.Person;
 public interface PersonDao {
 	public void save(Person p);
 
-	public Person getById(int id);
+	public Person getById(Long id);
 
 	public void delete(Person p);
